executors: Make setup simpler
Outside of our provided terraform modules, the setup of executors is very tedious and error prone today.
This PR tries to address that by adding as much validation of the env as possible, adding hopefully helpful resolutions, and removing some of the manual steps that we did in the VM install script.